Create README.md
Motivated professional seeking a career transition into the field of business analysis. Proficient in Excel, SQL, Python, Machine Learning, and data visualization tools such as Tableau and Microsoft Power BI. Currently excelling as a Seller Partner Support mentor at Amazon, where I have developed strong problem-solving and communication skills in a fast-paced, customer-centric environment. With a keen interest in data analysis and a drive to learn and grow in this field, I am eager to leverage my transferable skills and analytical mindset to contribute to the success of a data-driven organization. Committed to acquiring the necessary knowledge and expertise to excel as a business analyst and make a positive impact.